Adding a new file with name admin

+++ todos.py
@@ -77,11 +77,14 @@ def update_todo(user: user_depensy, db: db_depensy, todo_id: int, todo_request: 
     db.commit()
 
 @router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(db: db_depensy, todo_id: int):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+def delete_todo(user: user_depensy, db: db_depensy, todo_id: int):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed!!!')
+
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filetr(Todos.owner_id == user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail='TODO NOT FOUND!!!')
 
-    db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
+    db.query(models.Todos).filter(models.Todos.id == todo_id).filetr(Todos.owner_id == user.get('id')).delete()
     db.commit()
